pyproject.toml
[project]
name = "chatbot-ex-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-genai>=1.35.0",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
